requirements.txt
openenv-core>=0.2.3
openai>=2.7.2
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
